Validate CAM & DENM basic test cases

+++ scripts/etsi_its_patches/geonw_generate_makefile.bash
@@ -1,5 +1,6 @@
 #!/bin/bash
-set -x
+#set -e
+set -vx
 
 function f_exit {
     cd ${CURPWD}
@@ -84,31 +85,31 @@ then
     fi
 fi
 
-REFERENCES="LibCommon LibIts/Common LibIts/BTP LibIts/GeoNetworking LibIts/GeoNetworking LibIts/Ipv6OverGeoNetworking LibIts/Security LibIts/CAM LibIts/DENM"
+REFERENCES="LibCommon LibIts/Common LibIts/BTP LibIts/GeoNetworking LibIts/Ipv6OverGeoNetworking LibIts/Security LibIts/CAM LibIts/DENM"
 for i in ${REFERENCES}
 do
     # TTCN code
     for j in `find ${PATH_DEV_ITS}/src/$i/ttcn -type f -name "*.ttcn"`;
     do
-	ln -sf $j ../ttcn/`basename $j`
+	      ln -sf $j ../ttcn/`basename $j`
     done
     # Include source code
     files=`find ${PATH_DEV_ITS}/src/$i/include -type f`
     if [ "${files}" != " " ]
     then
-	for j in ${files};
-	do
-	    ln -sf $j ../include/`basename $j`
-	done
+	      for j in ${files};
+	      do
+	          ln -sf $j ../include/`basename $j`
+	      done
     fi
     # CC source code
     files=`find ${PATH_DEV_ITS}/src/$i/src -type f`
     if [ "${files}" != " " ]
     then
-	for j in ${files};
-	do
-	    ln -sf $j ../src/`basename $j`
-	done
+	      for j in ${files};
+	      do
+	          ln -sf $j ../src/`basename $j`
+	      done
     fi
 done
 
@@ -119,14 +120,13 @@ TTCN_FILES=`find .. -name '*.ttcn*'`
 if [ "${OSTYPE}" == "cygwin" ]
 then
     rm ../bin/*.exe ../lib/*.dll
-    compiler.exe -b -d -f -g -j -l -L -t -R -U none -x -X ${TTCN_FILES} ${ASN1_FILES} 2>&1 3>&1 | tee build.log
+    compiler.exe -d -e -f -g -j -l -L -t -R -U none -x -X ${TTCN_FILES} ${ASN1_FILES} 2>&1 3>&1 | tee build.log
     if [ "$?" == "1" ]
     then
         f_exit "Failed to compile ATS" 4
     fi
 else
-    compiler -d -f -g -l -L -t -R -U none -x -X ${TTCN_FILES} ${ASN1_FILES} 2>&1 3>&1 | tee build.log
-#    compiler -b -d -f -g -j -l -L -t -R -U none -x -X ${TTCN_FILES} ${ASN1_FILES} 2>&1 3>&1 | tee build.log
+    compiler -d -e -f -g -l -L -t -R -U none -x -X ${TTCN_FILES} ${ASN1_FILES} 2>&1 3>&1 | tee build.log
     if [ "$?" == "1" ]
     then 
         f_exit "Failed to generate ATS source code" 6
@@ -158,26 +158,39 @@ fi
 # Remove port skeletons to use src/<port skeletons>
 for i in `ls ../include/*.hh`
 do
-    rm ./`basename $i`
+    if [ -f ./`basename $i` ]
+    then
+	      rm ./`basename $i`
+    fi
 done
 for i in `ls ../src/*.cc`
 do
-    rm ./`basename $i`
+    if [ -f ./`basename $i` ]
+    then
+	      rm ./`basename $i`
+    fi
 done
+
+# Check if Makefile was generated
+if [ ! -f ./Makefile ]
+then
+    f_exit "Failed to generate ATS source code" 8
+fi
 
 # Patch ATS generated files
 #./bin/patch.bash 2>&1 3>&1 | tee --append build.log
 # Add compiler/linker options
+# -DASN_DISABLE_OER_SUPPORT is required for CAMCodec and DENMCodec
 if [ "$1" == "prof" ]
 then
-    CXXFLAGS_DEBUG_MODE='s/-Wall/-pg -Wall -std=c++11 -pthreads -fstack-check -fstack-protector/g'
+    CXXFLAGS_DEBUG_MODE='s/-Wall/-pg -Wall -std=c++11 -pthreads -fstack-check -fstack-protector -DASN_DISABLE_OER_SUPPORT/g'
     LDFLAGS_DEBUG_MODE='s/LDFLAGS = /LDFLAGS = -pg -pthread -fstack-check -fstack-protector/g'
 else
-    CXXFLAGS_DEBUG_MODE='s/-Wall/-ggdb -O0 -Wall -std=c++11 -pthread -fstack-check -fstack-protector/g'
+    CXXFLAGS_DEBUG_MODE='s/-Wall/-ggdb -O0 -Wall -std=c++11 -pthread -fstack-check -fstack-protector -DASN_DISABLE_OER_SUPPORT/g'
     LDFLAGS_DEBUG_MODE='s/LDFLAGS = /LDFLAGS = -g -pthread -fstack-check -fstack-protector/g'
 fi
-ADD_INCLUDE='/CPPFLAGS = /a\\CPPFLAGS += -I$(PATH_DEV_ITS)/include -I$(PATH_DEV_ITS)/include/asn1 -I$(PATH_DEV_ITS)/framework/include -I../include -I../../LibIts/Common/include -I../../LibIts/BTP/include -I../../LibIts/CAM/include -I../../LibIts/DENM/include -I$(HOME_INC) -I.'
-ADD_LIBRARIES='s/LINUX_LIBS = -lxml2/LINUX_LIBS = -lxml2 -lpcap -L$(PATH_DEV_ITS)\/lib -lItsAsn /g'
+ADD_INCLUDE='/CPPFLAGS = /a\\CPPFLAGS += -I/usr/local/share -I$(PATH_DEV_ITS)/include -I$(PATH_DEV_ITS)/include/asn1 -I$(PATH_DEV_ITS)/framework/include -I../include -I../../LibIts/Common/include -I../../LibIts/GeoNetworking/include -I$(HOME_INC) -I.'
+ADD_LIBRARIES='s/LINUX_LIBS = -lxml2/LINUX_LIBS = -lrt -lxml2 -lpcap -L$(PATH_DEV_ITS)\/lib -lItsAsn /g'
 sed --in-place "${CXXFLAGS_DEBUG_MODE}" ./Makefile 
 sed --in-place "${LDFLAGS_DEBUG_MODE}" ./Makefile
 sed --in-place "${ADD_INCLUDE}" ./Makefile
@@ -195,7 +208,7 @@ ADD_PORT='/PLATFORM = /aPORT=12000'
 sed --in-place "${ADD_PORT}" ./Makefile
 sed --in-place "${ADD_HOST}" ./Makefile
 ADD_RUN_LINE_1='$arun: all'
-ADD_RUN_LINE_2='$a\\t@$(PWD)/../bin/$(EXECUTABLE) $(HOST) $(PORT)'
+ADD_RUN_LINE_2='$a\\t@sudo LD_LIBRARY_PATH=$(LD_LIBRARY_PATH) $(PWD)/../bin/$(EXECUTABLE) $(HOST) $(PORT)'
 sed --in-place "${ADD_RUN_LINE_1}" ./Makefile 
 sed --in-place "${ADD_RUN_LINE_2}" ./Makefile 
 ADD_RUN_LINE_1='$arun_d: all'
@@ -203,7 +216,12 @@ ADD_RUN_LINE_2='$a\\t@gdb --args $(PWD)/../bin/$(EXECUTABLE) $(HOST) $(PORT)'
 sed --in-place "${ADD_RUN_LINE_1}" ./Makefile 
 sed --in-place "${ADD_RUN_LINE_2}" ./Makefile 
 ADD_RUN_LINE_1='$arun_v: all'
-ADD_RUN_LINE_2='$a\\t@valgrind -v --tool=memcheck --leak-check=yes --show-reachable=yes --track-fds=yes --run-cxx-freeres=yes $(PWD)/../bin/$(EXECUTABLE) $(HOST) $(PORT)'
+ADD_RUN_LINE_2='$a\\t@sudo LD_LIBRARY_PATH=$(LD_LIBRARY_PATH) valgrind -v --tool=memcheck --leak-check=yes --show-reachable=yes --track-fds=yes --run-cxx-freeres=yes $(PWD)/../bin/$(EXECUTABLE) $(HOST) $(PORT)'
+sed --in-place "${ADD_RUN_LINE_1}" ./Makefile 
+sed --in-place "${ADD_RUN_LINE_2}" ./Makefile 
+# Add gendoc entry
+ADD_RUN_LINE_1='$agendoc: ../docs/o2.cfg'
+ADD_RUN_LINE_2='$a\\tdoxygen ../docs/o2.cfg'
 sed --in-place "${ADD_RUN_LINE_1}" ./Makefile 
 sed --in-place "${ADD_RUN_LINE_2}" ./Makefile 
 
@@ -211,6 +229,7 @@ sed --in-place "${ADD_RUN_LINE_2}" ./Makefile
 make all 2>&1 3>&1 | tee --append build.log
 if [ "$?" == "1" ]
 then
-    f_exit "Failed to generate ATS source code" 8
+    f_exit "Failed to generate ATS source code" 9
 fi
+../bin/Ats${ATS_NAME} -v
 f_exit "Build done successfully" 0
